+++ packages/asc-ui/src/components/ButtonBar/ButtonBar.tsx
@@ -3,12 +3,12 @@ import ButtonBarStyle, { Props as ButtonBarStyleProps } from './ButtonBarStyle'
 
 export type Props = ButtonBarStyleProps
 
-const ButtonBar: React.FC<Props> = ({ children, ...otherProps }) => {
-  return <ButtonBarStyle {...otherProps}>{children}</ButtonBarStyle>
-}
-
-ButtonBar.defaultProps = {
-  padding: 5,
-}
+/**
+ *
+ * @deprecated
+ */
+const ButtonBar: React.FC<Props> = ({ children, ...otherProps }) => (
+  <ButtonBarStyle {...otherProps}>{children}</ButtonBarStyle>
+)
 
 export default ButtonBar