+++ packages/asc-ui/src/components/ButtonBar/ButtonBarStyle.ts
@@ -1,12 +1,9 @@
 import styled from '@datapunt/asc-core'
 
-export type Props = {
-  padding?: number
-}
+export type Props = {}
 
 const ButtonBarStyle = styled.div<Props>`
   display: flex;
-  ${({ padding }) => `padding: ${padding}px;`}
 
   & > * {
     margin-right: 5px;